+++ core/src/test/java/org/testifyproject/core/util/FileSystemUtilTest.java
@@ -19,6 +19,7 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Set;
 import static org.assertj.core.api.Assertions.assertThat;
 import org.junit.Before;
 import org.junit.Test;
@@ -112,6 +113,28 @@ public class FileSystemUtilTest {
 
         assertThat(result).isDirectory();
         assertThat(result.list()).isEmpty();
+    }
+
+    @Test
+    public void givenGlobForNonExistentFilesFindClasspathFilesShouldReturnEmptyList() {
+        String[] patterns = {"*.testext"};
+
+        Set<Path> result = sut.findClasspathFiles(patterns);
+
+        assertThat(result).isEmpty();
+    }
+
+    @Test
+    public void givenGlobForExistentFilesFindClasspathFilesShouldReturnList() throws IOException {
+        String pattern = String.format(
+                "**/{%s,%s}.class",
+                FileSystemUtil.class.getSimpleName(),
+                FileSystemUtilTest.class.getSimpleName()
+        );
+
+        Set<Path> result = sut.findClasspathFiles(new String[]{pattern});
+
+        assertThat(result).hasSize(2);
     }
 
 }